More modifications to eth contract
